requirements.txt
google-genai
telethon
psutil
python-dotenv
